% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSigma.R
\name{estimateSigma}
\alias{estimateSigma}
\title{Estimate the standard deviation of the intensities between two adjacent change-points}
\usage{
estimateSigma(Y, h = 10)
}
\arguments{
\item{Y}{the numeric vector of the intensities of markers}

\item{h}{the bandwidth of the local smoother}
}
\value{
This function estimates the standard deviation of the intensities between two adjacent change points
}
\description{
This function estimates the standard deviation between two adjacent change-points using a local smoother.
}
