% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrr.R
\docType{data}
\name{example.data.lrr}
\alias{example.data.lrr}
\title{example.data.lrr}
\format{A data frame with 10 individuals on 50000 markers}
\usage{
data(example.data.lrr)
}
\description{
An example data to demonstrate modSaRa2 processing multiple sequences
}
\details{
This example data is normalized measure of signal intensities for 10 sequences. Each column represents data for one individuals and each row represents data for one marker.
}
\examples{
# input the data
data(example.data.lrr)
data <- example.data.lrr
}
\keyword{datasets}
