% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{example.data.map}
\alias{example.data.map}
\title{example.data.map}
\format{A data frame with 50000 markers on the following three variables:
Name as SNP identifier;
Chr as chromosome ID;
Position as the chromosomal coordinates.}
\usage{
data(example.data.map)
}
\description{
An example map file to demonstrate modSaRa
}
\examples{
# input the data
data(example.data.map)
# get general information about the three variables
str(example.data.map)
}
\keyword{datasets}
