% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fInverse.R
\name{fInverse}
\alias{fInverse}
\title{Get the inverse cumulative distribution function of local min p-values}
\usage{
fInverse(n = 10000, h = 10, hh = 2 * h, precise = 10000, simT = 100)
}
\arguments{
\item{n}{the length of the data to simulate}

\item{h}{the bandwidth for the screening procedure. Defaults to 10}

\item{hh}{the bandwidth for the local minimum procedure}

\item{precise}{the precision of the approximation (the number of quantiles to use)}

\item{simT}{number of simulations}
}
\value{
the quantiles of the approximated inverse CDF
}
\description{
This function approximates the inverse cumulative distribution function (CDF) of the local min p-values via simulation.
}
