% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOneBIC.R
\name{getOneBIC}
\alias{getOneBIC}
\title{The modified Bayesian Information Criterion step to remove false positives in change-points}
\usage{
getOneBIC(x, cp, mod = FALSE)
}
\arguments{
\item{x}{the vector of the intensities of markers}

\item{cp}{the vector of the marker index of the identified change-points}
}
\value{
a list of change-points after filtering false positives
}
\description{
The modified Bayesian Information Criterion step to remove false positives in change-points
}
