% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli.single.R
\name{Likeli.single}
\alias{Likeli.single}
\title{Likeli.single}
\usage{
Likeli.single(lrr, laf)
}
\arguments{
\item{lrr}{the matrix of the log R ratio intensities. Each column describes a single sample or sequence and each row describes a single marker}

\item{baf}{the matrix of the B Allele Frequency intensities. Each column describes a single sample or sequence and each row describes a single marker}
}
\value{
X the calculated copy number estimate

L the vector of likelihood of being each copy number state

L_geno the vector of likelihood of being each genotype
}
\description{
This function integrates the LRR and BAF intensities to calculate the likelihood of being each copy number state for a position in a sequence
}
