% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMax.R
\name{localMax}
\alias{localMax}
\title{Get the local maximizers of local diagnostic function}
\usage{
localMax(y, span = 5)
}
\arguments{
\item{y}{the list of local diagnostic values within a small neighborhood [x-h, x+h]}

\item{span}{the bandwidth to find local Maximizer of local diagnostic function}
}
\value{
numeric vector of position index for all the local maximizers of local diagnostic function
}
\description{
This function finds the local maximizers of local diagnostic function |D(x,h )| and returns the value of position index for all the local maximizers.
}
\details{
Local maximizer is defined as follows. For any number x, the interval (x-h, x+h) is called the h-neighborhood of x. And, x is an h-local maximizer of function f(.) if reaches the maximum at x within the h-neighborhood at x. In other words, f(x)>=f(x') for all x' in (x-h, x+h).
}
